Polished the "A" filter, fully implemented "Bare" display mode
Also began work on the sorting algorithm"O" as well as testing everything in the test file. Some flow optimization have also been implemented in the "A" filter (the conditionals of the arguments were a bit wrong), also fully implemented the opposite arguments of the "A" filter.
Also made the total file and directories counter in the footer of the standard display.

TODO:
* Finish the headers and footers of the standard display
* Sorting
* Get ADSReader working