Memory optimizations and sorting by extension
I've now made sure that the ArrayList is converted into a normal array as soon as we know the final number of paths, which is after the filtering. I've replaced all the references after the filters from ArrayList into arrays. I've also added a sorting algorithm for sorting paths by their extension

TODO:
* Add capital letter synonym arguments for the "A" and "O" options (perhaps JOpt has this feature? Need to research that.
* Implement a reverse sorting by extension algorithm
* Implement all the other sorting algorithms