Implemented "X" and filtered out non-path wildcards
It turns out that the Windows command shell actually automatically takes any path wildcard and converts it into an array of all the paths that fits the wildcard. There is nothing I can do in Java to differentiate between an actual user-inputted argument and a Windows parsed wildcard path. Since the DirectoryStream class throws an exception when handling a path that isn't a directory, I was forced to simply ignore non-directory paths if they are given as arguments. I'd like to offer a cleaner solution, but, unfortunately, Java isn't the proper tool for this.

TODO:
* Debug why junction dates in the sort function are really their resolved path dates
* Figure out how to pipe any output to Windows "more"